+++ src/main/java/uk/co/qa/domain/Programme.java
@@ -4,19 +4,19 @@ import java.util.Objects;
 
 public class Programme {
 
-    private long id;
+    private int id;
     private String name;
 
-    public Programme(long id, String name) {
+    public Programme(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
